Revert "Restore add channel header empty states"
