+++ webapp/channels/src/components/channel_header/channel_header_text.test.tsx
@@ -91,4 +91,55 @@ describe('ChannelHeaderText', () => {
 
         expect(container.childNodes.length).toBe(0);
     });
+
+    test('should return null for public channels without header regardless of permissions', () => {
+        const channel = TestHelper.getChannelMock({
+            type: 'O',
+            header: '',
+        });
+
+        const state = {
+            entities: {
+                channels: {
+                    myMembers: {
+                        [channel.id]: {channel_id: channel.id, roles: 'channel_role'},
+                    },
+                    roles: {
+                        [channel.id]: new Set(['channel_role']),
+                    },
+                },
+                teams: {
+                    myMembers: {
+                        [defaultTeamId]: {team_id: defaultTeamId, roles: 'team_role'},
+                    },
+                },
+                users: {
+                    currentUserId: 'user_id',
+                    profiles: {
+                        user_id: {
+                            id: 'user_id',
+                            roles: 'system_role',
+                        },
+                    },
+                },
+                roles: {
+                    roles: {
+                        system_role: {permissions: ['test_system_permission']},
+                        team_role: {permissions: ['test_team_permission']},
+                        channel_role: {permissions: ['manage_public_channel_properties']},
+                    },
+                },
+            },
+        };
+
+        const {container} = renderWithContext(
+            <ChannelHeaderText
+                teamId={defaultTeamId}
+                channel={channel}
+            />,
+            state,
+        );
+
+        expect(container.childNodes.length).toBe(0);
+    });
 });
