+++ webapp/channels/src/components/channel_info_rhs/about_area_channel.tsx
@@ -122,13 +122,13 @@ const AboutAreaChannel = ({channel, canEditChannelProperties, actions}: Props) =
                 </ChannelPurpose>
             )}
 
-            {(channel.header || canEditChannelProperties) && (
+            {channel.header && (
                 <ChannelHeader>
                     <ChannelDescriptionHeading>
                         {formatMessage({id: 'channel_info_rhs.about_area.channel_header.heading', defaultMessage: 'Channel Header'})}
                     </ChannelDescriptionHeading>
                     <EditableArea
-                        content={channel.header && (
+                        content={(
                             <LineLimiter
                                 maxLines={4}
                                 lineHeight={20}
@@ -141,7 +141,6 @@ const AboutAreaChannel = ({channel, canEditChannelProperties, actions}: Props) =
                         editable={canEditChannelProperties}
                         onEdit={actions.editChannelHeader}
                         editTooltip={formatMessage({id: 'channel_info_rhs.about_area.edit_channel_header', defaultMessage: 'Edit channel header'})}
-                        emptyLabel={formatMessage({id: 'channel_info_rhs.about_area.add_channel_header', defaultMessage: 'Add a channel header'})}
                     />
                 </ChannelHeader>
             )}
