+++ webapp/channels/src/components/channel_info_rhs/about_area_gm.test.tsx
@@ -172,19 +172,4 @@ describe('channel_info_rhs/about_area_gm', () => {
 
         expect(screen.getByText('my channel header')).toBeInTheDocument();
     });
-
-    test('should display add channel header empty state when header is empty', () => {
-        renderWithContext(
-            <AboutAreaGM
-                {...defaultProps}
-                channel={{
-                    ...defaultProps.channel,
-                    header: '',
-                }}
-            />,
-            initialState,
-        );
-
-        expect(screen.getByText('Add a channel header')).toBeInTheDocument();
-    });
 });
